+++ apps/skde/src/components/IndicatorTable/Indicatortable/IndicatorSection.tsx
