+++ apps/skde/src/components/IndicatorTable/Indicatortable/IndicatorTableStyles.tsx
